move conditions to new pkg
